requirements.txt
networkx==3.1
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
python-dotenv==1.0.0
mysql-connector-python==8.2.0
sqlalchemy
networkx==3.1
